Fix Compose completion outside of functions
This change fixes an issue where the Compose insertion handler would be
triggered even in locations where the completion of methods is not the
right resolution.
For imports, the Compose insertion handler would treat it as a regular
method call insertion and add the parameters, breaking the user code.
This changes the ComposeCompletionContributor so it only triggers when
the insertion is being done inside a KtFunction.

Fixes: 153769933
Test: ComposeCompletionContributorTest
Change-Id: I9a30a294f92c5d3ebb517d17b312a49b1ee1e24a