Delete Items From History Using Del Key
Ideally, we should also have a x icon on each item that when clicked will delete the item from history, but I couldn't find an easy way to do this.

It's easy to render an icon but the list doesn't forward events to it.

I had to remove a couple of tests. The change from:

  JBPopupFactory.getInstance().createPopupChooserBuilder()

To:
  PopupChooserBuilder()

Makes it hard to test.

Bug: 191880898
Test: Manually
Change-Id: I8c0e6a509b241aee9ce9ec6c2b24d372d3e11d22