[UTP AGP] added tasks to TestApplication
Added managed device and device group tasks to TestApplication. This
should fix an issue where test applications do not
run when the managed device check is run.

This does not modify existing tasks, but rather creates a new set of
tasks that are added dependencies of the following tasks:

<managedDeviceName>Check
<managedDeviceGroupName>groupCheck
allDevicesCheck

This is similar to how these test tasks are currently done for
connectedCheck with the TestApplicationTaskManager

Bug: 195581855
Test: Ran on local project with a test-only module and ensured tasks were created appropriately. Added an instrumentation test to ensure tasks are run.
Change-Id: I597e5d8da2e293411f53ac17bc8a281e55e9c242