[AUA] Address slow operations on EDT
We are already careful to parse the project to build the Build models
in a background thread.  However, that does not (quite) cover all
potentially-slow operations; component refactoring processors, such as
the Gradle Version processor, might inspect the Psi directly (for
example, looking at the Gradle wrapper properties file).  Compute and
cache the is no-op values for each processor in ensureParsedModels().

Bug: 218987487
Test: none
Change-Id: Ie4e85dbc5a3d1bb0000b94f25b4d767c3fabdf9b