Proactively fix a UI test
The test assumes that configuring Kotlin in a project will introduce
the text `ext.kotlin_version` to the root Gradle build file, but after
a forthcoming change that will no longer be true.  Relax the text
search and manipulation just enough to pass in the new world as well
as the old.

Bug: 180411529
Bug: 217805224
Test: existing
Change-Id: I18a3736ed0bd3bf97711e89d287fc141468e5c11