Make login be a hard requirement to all functionality
With the current state of matters, being logged out is modeled per
module and is independent of us discovering firebase configurations for
an app. The way it will manifest in our UI is that the tool window will
have the app selector and all tabs in it will have a zero-state "Please
login" UI in it.

This change proposes to change it and make "being logged-in" a global
"error" state where:

* The tool window has only one "content" with the login zero state.
* We don't use any resources like discovering apps and
  google-services.json files until the user gets logged in.

Bug: b/222718017
Test: manual since this is a temporary implementation.
Change-Id: I8ba59fbc7f828839a3de29e6786e7ff9181b3f51