Add VarHandle benchmarks in the form of a test.
The benchmarks are autogenerated with a python script in order to avoid
copy-paste due to a large number of method variations. Currenly the
script generates 138 benchmarks in separate files for different methods.

The benchmarks include a few methods from Unsafe and Reflect to compare
them with VarHandle equivalents.

These should be in golem, but it is blocked on a toolchain update.
Add benchmarks as a test for now to prevent bitrotting.

The test is disabled for `--jvm` option.

Bug: 71781600
Test: art/test/run-test --host 2239-varhandle-perf
Test: art/test/run-test --host 2239-varhandle-perf --64
Test: art/test/run-test 2239-varhandle-perf
Test: art/test/run-test 2239-varhandle-perf --64
Change-Id: Ibe6b253d547066cbba4312b6c20e8c76f62be0e5