Autoupdate of 2 submodules
*tools/adt/idea updated to ec434e72254ee1fa06b4bb4d262cb5f180a2aaad
222095229: Move VersionChecks into API package to open up visibility

(Tracks changes in the lint module.)

Fixes: 222095229
Test: Existing unit tests
Change-Id: Iad9121a81658ba0a7250f75a001f68810aeba27b

*tools/base updated to 945f5bc070a1d5255948646d253127edf0a0387a
222095229: Move VersionChecks into API package to open up visibility

Leaves the old signatures in place temporarily since they're used from
places like androidx.

Despite all the diffs, this basically just moves a bunch of existing
code from the lint-checks module to the lint-api module; the
VersionChecks file as a whole (which isn't showing up as moved since we
kept the original in order to forward method signatures), plus
associated parts from the ApiDetector and the SdkIntDetector. It also
registers a new place-holder issue used exclusively for partial state
persistence, since we cannot stash the is-version-check-method data
using the SkdIntDetector issue which is not available from within
lint-api.

Fixes: 222095229
Test: Existing unit tests
Change-Id: I610915bb85abbccba0e12f31a38a65153acead70