Initial AppQuality toolbar UI implementation.
We are using placeholders for the different  states, logged in or logged out, and sections,
but we create tabs based on the data from the Insights service.

This is a rework of ag/17059730 since it got reverted because of breaking changes.

Logged out screenshot @ https://screenshot.googleplex.com/7dLJDgxiFxQstUT
Logged in screenshot @ https://screenshot.googleplex.com/8drotLpBqZqpLeX

Test: manual
Bug: b/222303367
Change-Id: I41b051355792fd01b78c41c6a709d416abe4a440