Autoupdate of 2 submodules
*tools/analytics-library updated to 191f1780947924df46ff4f66c9fa95f9f5f5d08d
added enum values for added tasks.

Test: Existing
Bug: N/A
Change-Id: Ibecbf061c84968fa36e7964128dd7165ad989c18

*tools/base updated to f3473c240082733842cb2970bc0abab5bf2ea7f9
first implementation of the fused libraries plugin.

So far, it only does a casual class merging and use normal jar task
to bundle things up. This will be refined in subsequent CL.

Plugin itself sets up various configurations needed to properly
support consuming libraries that are merged and continue having
the other libraries as dependencies.

Minimal extension is registered that allow setting the desired
namespace for the new fused library (not used yet by tasks).

publication is suppored although ran into issues with
configuration caching (to be resolved in subsequent CL if it
is not a Gradle issue).

Test: Added some integration tests and unit tests.
Bug: 62121508
Change-Id: I3b43c133feceaaf2ea3a0553fbb19ca6eb7312df