Adapt Preview caches to Multipreview
Some Preview related caches were based on modification trackers
considering only the file where a given Composable is declared.
But with Multipreview, it is possible that modifications to other
files implicate a change in the Previews, and as a consequence,
the need of marking some caches as stale.

This change modifies these Preview caches to be marked as stale on
any modification to any Kotlin and or Java file.

Fixes: 223208889
Test: manually (automated to be added later, tracked by b/223425524)
Change-Id: Ief566b85d1bab71e60d6d18c5e903a93b5162359