Remove superfluous tooltip from component tree
The tooltip in the component tree is superfluous because the tree
supports an ExpandableItemsHandler for any text that is partly
hidden when the tree panel is too narrow.

Only when a node has: id, tagName, and text do we show a tooltip.
Because in this case the tree renderer will not show the tagName,
only id and text.

Fixes: 223245502
Test: Unit tests updated.
Change-Id: I10da9bd82a840922a300b8a901254fe300161c6f