Fix NPE in Preview caused by incorrect initialization order
ComposePreviewRepresantation sets a callback function when creating
its 'previewFreshnessTracker', and that callback uses variables that
were being declared later in the initialization process. This was
causing some NPE as the callback could be executed before the whole
ComposePreviewRepresentation is initialized.

This change fixes the problem by modifying the declaration order of
the involved variables.

Fixes: 225043299
Test: manually by following the steps described in the bug.
Change-Id: Id1b955619589ae325c1d9902be5a70707c665f7c