Autoupdate of 2 submodules
*prebuilts/r8 updated to a305b4818c28494958c31c8e65ba5f6b657de438
Update D8 R8 to 3.3.19-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.3.19-dev
Test: ./gradlew check
Bug:

Change-Id: I965d1e6808386aad5d2dc41df508f5c17654dbc2

*tools/base updated to 69b84fabf16521c47855641fa8dcd12030aa8fac
Remove unused r8 imports

I assume these were auto added due to it ->

Test: AGP build
Bug: 225131995
Change-Id: I0a5bff627ae69ab46804e5da214e1f74c4db7325