Test that user 0's directories are encrypted
Add a test case which tests that all of user 0's directories that should
be encrypted actually are, and that user 0's CE and DE keys differ.

This does extend the scope of vts_kernel_encryption_test a bit beyond
its original scope, which was to just test the kernel's interface.  But
this is an important test case, and there doesn't seem to be a better
place for it.  (CTS doesn't have permission to do this test case.)
Also, the existing test cases already make assumptions about /data.

Bug: 164488924
Test: atest vts_kernel_encryption_test  # on Cuttlefish
Change-Id: I7710a58a9702eb16a3f3c2fd58ecfbad7c04f68b