Update D8 R8 to 3.3.22-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/3.3.22-dev
Test: ./gradlew check
Bug: b/225393424

Change-Id: I163dc1052a9d305377c8adbcc9a2fec0925c0d67