[BA] Fix UI on font/colors scheme change
Solve this by listening for the corresponding LAF topic and recreate
Build Analyzer UI completely on the event of these settings change.
Alternatively it is also possible to make each element respond to
the theme changes properly but overall it is easier to just recreate
the whole ui hierarchy as in this case we are 100% sure nothing is
cached somewhere resulting in the inconsistent change.
Since such change event is supposed to be extremely rare I believe
it is not a problem to recreate UI in this case.

Only recreate the UI elements tree, keep the model of the pages behind
intact. This makes BA UI to generally keep the state before the settings
change (e.g. opened pages).

Bug: 218431929
Test: manually verified
Change-Id: I2e3c83216fee1b54c66103e2cf4b534543f0d895