Remove deprecated Label() argument
relative_to_caller_repository is removed in Bazel 5.0, and is advised
not to be used in 4.2 documentation.

Running the command below, before and after, there appears to be no
differences to GenProto actions (only location this gets used).

```
bazel aquery 'mnemonic("GenProto", //tools/...)'
```

Produces the same result regardless of this argument.

Bug: 220902606
Test: n/a
Change-Id: I04abe657e04e0c1c38aabbdde3a3886279853b50