Autoupdate of 3 submodules
*tools/adt/idea updated to af82981ea1bc147ce2de0caf374667b3830e6f0f
Add more test coverage for Unit test execution

We need to know more about how the unit test execution is rendered on
the UI in AS. The way to know about this is by checking which component
handles the test events rendering on AS.
This change adds a test to verify that the different test run
configurations are rendered as expected by the
GradleTestsExecutionConsoleManager.

Bug: 220143136
Test: this
Change-Id: Ie92e5d9217a7444351538afadd4a708adf9366bc

*tools/idea updated to 2faee9f8da009700cd45f3edc7a5ed431da7ba98
Use TestData to determine if a task is a test

This is a temporary workaround for Android Studio in order to support
rendering the test directory Run Configurations results on the test
output window.
The long-term plan is for Jetbrains to fix it:
https://youtrack.jetbrains.com/issue/IDEA-289683

Bug: 217700611
Test: AndroidGradleConfigurationProducersTest
Change-Id: I3e64fd4f1d1ef8eff31115983644e96686486d1d

*prebuilts/studio/intellij-sdk updated to 846fe2b3d4e1bdb2bb54e4b5b0afb6a38342e246
Update AI-213 after commit 2faee9f8

Change-Id: I3e64fd4f1d1ef8eff31115983644e96686486d1d