Remove timeout-and-destroy behavior from exec()
This change reverts the timeout-and-destroy behavior added to exec() in
aosp/1728390, and instead adds it to a new execWithTimeout() method.
This is intended to avoid deadlocks caused by Traceur reading a Perfetto
process's input stream after waiting for its termination.

Calls to exec() that direct stdout to Logger have been replaced with
execWithTimeout().

Bug: 226969096
Test: - Recorded normal and long traces to check that traces are still
        collected correctly.
      - Ran TraceurAppTests#testSuccessfulTracing multiple times to
	check that the stdio deadlock no longer occurs.
Change-Id: I8cc7d26246083dbad93fe0d9dfac520e27dd6586