Autoupdate of 2 submodules
*prebuilts/studio/jdk updated to 866dad10b3cf9a966aed7cc312e77a54346198f1
Add BUILD file for @studio_jdk repo

The BUILD only contains langtools as this is the only target that needs
remapping for the @blaze repository.

Bug: 187294442
Test: n/a
Change-Id: I8fb52990f92780c3b6b65d082434ffe3ace8b436

*tools/base updated to ab2f4544168d8bf6ed5b87b46bef4a91e38de06c
Remove JAVA_HOME

This removes the default local java targets in the @local_jdk
repository. A @studio_jdk repository is added to satisfy the @bazel
repository which depends on @local_jdk//:langtools

Bug: 187294442
Test: n/a
Change-Id: I55725c022a3b85c175b457598020a53f524844c4