Autoupdate of 2 submodules
*tools/adt/idea updated to be2cdc92f720b147ace6d88fcf1e60315c7eddb1
Change Methodology of FakeAdbHostServices.trackDevices() to Use a Channel

This makes it possible to control better, for example, one can close() the channel causing an exception to test error conditions.

Bug: 227905221
Bug: 191880898
Test: Updated
Change-Id: I2f8e6577e257e309b51a2a51eefbb2ce951fffca

*tools/base updated to 505ebffeea00a9008ff376278397d52496d74aa3
Change Test Methodology of trackDevices() to Use a Channel

This makes it possible to control better, for example, one can close() the channel causing an exception to test error conditions.

Bug: 227905221
Bug: 191880898
Test: Updated
Change-Id: I2f8e6577e257e309b51a2a51eefbb2ce951fffca