Autoupdate of 2 submodules
*tools/adt/idea updated to f00ff604c0c51057fb754cd8c42c5e19e3641950
Fix memory leak of RenderErrorContributor

RenderErrorContributor has a reference to ResultResults, which has
reference to the BufferedImage, but RenderErrorContributor doesn't need
to have the information of BufferedImage. Read the needed attributes
instead of referencing RenderResult in RenderErrorContributor.

Fixes: 217568626
Test: N/A
Change-Id: I0753e2ba381ba95d5ba10d538d43dbfc5501839e

*tools/vendor/google3/blaze updated to dc5b6a81233e0aca296f7095fb02aa1d776981bd
Update report issue function in RenderErrorLoggingContributor

Update the function call for the change in RenderErrorContributor

Fixes: 217568626
Test: N/A
Change-Id: I4d654bce879641c7d9284de01f849bf50d4c5006