Autoupdate of 2 submodules
*prebuilts/tools updated to a7fa21d57883411805de5fbd7fc8640035254f3e
Replace bazel-real with bazelisk

Bazelisk 1.11.0
https://github.com/bazelbuild/bazelisk/releases/tag/v1.11.0

Bazelisk makes it easier to switch between Bazel versions and avoid
checking in large bazel binaries.

Bug: n/a
Test: n/a
Change-Id: Id85776cdf95c4e6caf2dcbd24615d948edd0c1a1

*tools/base updated to e5a3b2215adab55d80af7a8f61794b157b2cf214
Use bazelisk and Bazel 5.1.1

https://github.com/bazelbuild/bazel/releases/tag/5.1.1

This release contains an important fix for remote execution:
https://github.com/bazelbuild/bazel/pull/15181

The PATH environment variable needs to be included for linux, or else
`bazel run` commands will fail to find `bash` in order to run
executables. A future change will remove explicit passing of environment
variables.

Bazelisk expects to find the LocalAppData environment variable for where
to store Bazel downloads. This had to be added for Bazelisk to work.

Bug: n/a
Test: n/a
Change-Id: I5ff925eef37c3648ee10e49f90ae2b846266a7a5