Inherit language level from the project
The .iml files for ASwB modules were explicitly setting
the java source level to 1.8. This is an issue since we are
moving to 2021.1 which requires 1.11.

But there is no reason to be using 1.8 for ASwB as Google3
has moved to 1.11 already. This CL changes all the iml files
to inherit the language level setting from the project.

Separately, I've confirmed that explicitly setting the language
level to 1.11 for these modules also works, provided the project
itself is updated to target 1.11, and also includes an
additional parameter to javac: https://paste.googleplex.com/4619593923428352?raw

Bug: N/A
Test: compilation succeeds
Change-Id: I2f7f33aa3a6d8b63d519cc9ba81cfca033fd1a04