Remove LARGE_MACHINE from targets
The targets using larger VMs are no longer the majority of critical path
targets. To simplify the pool setup, we are going to move all tests back
to the default pool.

The shard count for two split targets is increased since these targets
start to time out. Having an extra shard for this target is
not a problem, since using a single pool should bring better sharing of
machines.

Bug: 165823108
Test: n/a
Change-Id: Ie164626d91ca80a214b8ec5703fc23ff84a517c6