Catch IO exceptions when making rpc calls
This is just the initial naive fix to avoid forever
loading panels.

The original issue might be from how we handle Google
OAuth "invalid_grant" in studio, not just in our own
crashlytics gRPC clients.

For what we show to the users when running into these
unpleasant exceptions, it would be handled in the other
future CLs.

Bug: 229279445
Test: n/a
Change-Id: Ib0aa24fb8adebaaf7231b0793bc5f558d9340f3a