Autoupdate of 2 submodules
*tools/adt/idea updated to d4e85963919b2e68642999b8ec94d974113cf938
Update LE to handle incompatible changes

This changes causes LE to enter into an error state upon receiving
errors from the device regarding incompatible changes. An error string
detailing the message is displayed in the LE indicator, and LE does not
compile or push additional changes until the app is re-run.

Bug: 228632150
Test: Manual
Change-Id: I27447bb95d52641d6f9dccc63a84427199d1ca7e

*tools/base updated to b2fbb616af7872cdececa50905672e05a5b2ef0f
Add bytecode validation to LE runtime

When class updates are received, LE will ensure that certain conditions
are met before adding the classes/bytecode to the LE context. If
conditions are not met, an error object is returned detailing the issue.

Current checks:
 - Cannot add/remove methods (covers renaming, changing signatures,
   changing return types)
 - Cannot add classes

If any class in the update fails validation, none of the classes are
added to the context. If the classes have yet to be primed, failing
validation will also prevent priming.

Bug: 228632150
Test: BytecodeValidatorTest
Change-Id: Ib9976c94f626aa9fa7dddbd25df45fa5f610c938