Autoupdate of 3 submodules
*tools/data-binding updated to 943f692e4c7450ec19483a6d742233f6f9c9e08b
Compile data binding runtime using Gradle AGP

As this makes the path to compiling the runtime somewhat more
monolithic, this might start showing up in the critical path more often.

If that becomes a problem, we can investigate compiling the data binding
runtime for tests using a fixed version of AGP, and then having a
separate test that validates that the produced library is the same as is
produced by running the current AGP at head, with the option to
temporarily switch back if there are changes.

Bug: 143864616
Test: Migrating tests to use our production build
Change-Id: Ic5d35a8a2a5f9bd6bd37fb98084dc61534afd432

*tools/base updated to d78383b793c046892249601ae73a829ca538c90b
Switch real-world-integration-test to Gradle AGP

Bug: 143864616
Test: Included
Change-Id: I0d1848eb9fc9318c15408eb90814702f3b2d4787

*tools/vendor/google updated to d188f3f9cc34c244d3c5dda17ce9002205b01316
Switch real-world-integration-test to Gradle AGP

Bug: 143864616
Test: Included
Change-Id: I810f3fddb5395d4021ec439de839ff5325ac6125