Autoupdate of 2 submodules
*tools/adt/idea updated to a50a0a16c222fb335083faf74f9cd3ca75557820
[AUA] Introduce an error state into the AUA UX flow

This error state is used to display a diagnostic message (and disable
various graphical controls) when an internal error occurs; the
diagnostic message encourages the user to revert to a known-good state
before attempting to do anything else.

Introduce a new boolean to control whether the drop-down version combo
is active: it should be active in the previously-existing non-loading
states, but should be inactive if we have just hit an internal error.

Re-use the newly-introduced logic to disable the (empty) tree view of
Refactoring Processors in the AllDone state.

Bug: 227793240
Test: manual
Change-Id: Ifeef1f1c93f161a3f22e41a6bc8841e30d131a10

*tools/analytics-library updated to 5daa4976c801ba2a9371a93132fa98343063ef78
Update studio_stats.proto

Logs-Approval: cl/444416553
Logs-Approval: cl/444547714
Bug: 225388031
Bug: 227759559
Test: in accompanying CL (for Upgrade Assistant INTERNAL_ERROR)
Change-Id: I21e2c1521bebf2e14d4b854ce6f48cff79b3690e