Fix local process name prefix checking bug
When attaching to an Android app, we try to find which module in the
user's project might have produced that app. To do this, we compare
the package name of the module and the name of the process we are
attaching to.

The package name becomes a prefix of the process name when the app uses
local processes.

For instance:
  Package name of: com.example.myapp
  Local process name: com.example.myapp.localprocessname

To handle local processes, we use a prefix check.

However, there can be multiple modules in the project, where one is the
prefix of the other.

For instance:
  Module 1: app  (com.example.myapplication)
  Module 2: app2 (com.example.myapplication2)

In this case, (regardless of whether local processes are used) when we
try to attach to "app2", we accidentally match and use the "app" module,
because "com.example.myapplication" is a prefix of "com.example.myapplication2",

The fix in this change is to perform the prefix check against
package name + "." string.

Bug: n/a
Test: Manual; create above project layout and attach to app2
Change-Id: Ia7ae7fa847155d2143e14fc26a8099c12cb2c1ff