Autoupdate of 2 submodules
*tools/adt/idea updated to 9039b69335ab3ec329c43994706d76ea8a3679d9
Improve lint's Project#dependsOn lookup

IDE-side tweaks to adjust to the changes to lint base
changes for handling project lookup.

Test: AndroidLintTest (updated)
Bug: Not filed
Change-Id: Ia48a2c0afa2ebb5bd8b6f4e72b6a78ea6b5bfa3e

*tools/base updated to e66435200926d86e879f8c9ad3717153c075966b
Improve lint's Project#dependsOn lookup

First, this CL updates the project dependency lookup for AndroidX
appcompat to use class lookup as a fallback, which should be more
accurate for some build systems where we have incomplete dependency
information.

It also switches out from looking for ActionBarActivity to
AppCompatActivity, since ActionBarActivity was removed in later versions
of the pre-AndroidX support library and was replaced in favor of
AppCompatActivity.

This CL does some other cleanup as around this as well, such as getting
rid of old support-v4 lookups which were not being used, and some
subclasses of dependsOn which were not correctly handling lookup for
arbitrary dependencies.

Finally, it fixes a typo in the typo check -- which let to the
following very intermittent failure:
    java.io.FileNotFoundException: /Users/tnorbye/.android/cache/typos-en.txt-2.bin.1886371917 (Operation not permitted)
        at java.base/java.io.FileOutputStream.open0(Native Method)
        at java.base/java.io.FileOutputStream.open(FileOutputStream.java:298)
        at java.base/java.io.FileOutputStream.<init>(FileOutputStream.java:237)
        at com.google.common.io.Files$FileByteSink.openStream(Files.java:193)
        at com.google.common.io.Files$FileByteSink.openStream(Files.java:181)
        at com.google.common.io.ByteSink.write(ByteSink.java:101)
        at com.android.tools.lint.checks.TypoLookup$Companion.writeDatabase(TypoLookup.kt:569)

Test: Existing unit tests were updated
Bug: Not filed (but there have been bugs in google3 in the past where
     lint was incorrectly inferring whether appcompat was being used.)
Change-Id: I9ce43d466fbb00ad809cbc1fa397948f79ea0b29