230015328: Remove Unused Resources Refactor deleting all layout files in project
Lint will analyze its dependent library projects before analyzing its
own sources. However, the beforeCheckEachProject-hook was invoked
immediately, before processing the dependencies. This means that when
the main project is analyzed, the last beforeCheckEachProject it
observed may in fact not be from the current project but the last
dependent library project. This broke the unused resource detector which
was tracking whether to capture binding classes for the layout resources
in the current project.

Fixes: 230015328
Test: Unit test included
Change-Id: I8d866dd058b0976cc63a5ce80e64d983b5e8f2a0