[AUA] Disable activation of Build Window from tool window
If we are running from our own tool window, now that we treat Sync
failure as a state for us to handle within the Upgrade Assistant, we
need to disable the automatic activation of the Build output window if
the sync / refreshProject fails.  Use the new field in the
GradleSyncInvoker Request, making sure that the default state (for
when running outside the tool window, for example from a forced
upgrade dialog) is to show the build window on failure.

Bug: 227793240
Test: manual
Change-Id: I52963442e044528246cfc02c12367ef549d83176