Change MultiPreview DFS to start collecting metrics
This change modifies the MultiPreview DFS to allow its callers
to get all MultiPreview nodes, and not just PreviewElements.

Also, each time the DFS is executed, all these MultiPreview nodes
are used to collect MultiPreview metrics, making sure to avoid
logging the MultiPreview graph of a given file when the graph
hasn't change since the previous DFS executed for that file.

Fixes: 221276627
Test: MultiPreviewUsageTrackerTest

Change-Id: Ib7599669ccf5764bf8a174c900c04027976ab1a5