Autoupdate of 2 submodules
*prebuilts/tools updated to 71516083dce2f324a6600efbf0a4a4233aa0d21c
Update the compose-ui-inspection jar

From: Build 8507455

Bug: 229657989
Test: N/A
Change-Id: I1e0c23c7a081e8369af8313edd0ab905ca67abf5

*tools/adt/idea updated to 96455c89d6c8b0619a4ff4314591446c7a16f32f
Take advantage of delayes parameter extraction

Send "delay_parameter_extractions" with the update settings command.
If the compose inspector agent understands this option, we will
start to receive an anchorHash for each ComposeViewNode.
(and an anchor hash is also included in references for drilling into
 parameter details.)

We will send the anchor hash in the commands for getting parameters
and parameter details.

Test that the:
- "delay_parameter_extractions" option is specified
- the anchor hash is specified in get parameter commands
- the anchor hash is specified in references
- getComposables has `extractAllParameters` set correctly

Fixes: 229657989
Test: Update existing tests
Change-Id: I58113b53a4179301f1f41a52dc24f3947265a53a