Mark the NDK plugin as optional for ASwB
The NDK plugin is actually optional for ASwB.
This is correctly done in Google3 because
the plugin is built using Blaze and we use
the `optional_plugin_xml` target to indicate
this. But in the development set up in repo,
it is not marked as optional, which can lead
to ASwB plugin getting disabled if NDK fails
to load. To avoid this, this CL marks the NDK
plugin as optional even during ASwB development.

Bug: N/A
Test: ASwB plugin loads regardless of whether
NDK is enabled or not

Change-Id: Ifc75ed7eab72c2510320fdd51f8e5a0f638c3c72