Fix version dropdown updates
This changes makes use of StateFlows to keep track of the UI state of
the dropdown, and create a new tree everytime the dropdown is summoned.

Previous we maintained a single tree and tried to update it upon
collecting a new state. However, when the dropdown is already opened, we
ignored the new state and lost it forever.

This change fixes that by conforming how the dropdown UI is updated to
existing dropdowns in Studio, which is to update the next time the
dropdown is opened. (as opposed to updating the UI from under the feet
of the user).

Test: manual
Fixes: 231457596
Change-Id: I88d59235970406e2f671d2ee3c0d391b4a918192