Autoupdate of 2 submodules
*tools/base updated to e7d29ae261cf7580c372077a7a368d63446c0b2d
Add and test Abseil C++ project

Adding this because it's a good Prefab test bed. It has a relatively large
number of targets, and install configuration, and unittests.

Bug: 168994860
Test: Is test
Change-Id: Ia922ca6d3406ddc095605f0588a58ba440d7bfe5

*prebuilts/studio/buildbenchmarks updated to 6298c94b5709c73df8043612cf505aab1a9741f5
Add and test Abseil C++ project

Adding this because it's a good Prefab test bed. It has a relatively large
number of targets, and install configuration, and unittests.

Bug: 168994860
Test: Is test
Change-Id: I5a9feb2796c24f3399d0ad8abd54a7e3d3f4b120