Autoupdate of 2 submodules
*prebuilts/r8 updated to e41c8576ae0559f5d99410f495de8c5631629280
Update D8 R8 to 4.0.5-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/4.0.5-dev
Test: ./gradlew check
Bug:

Change-Id: I23efb75c820d752d2d980ea315397c13874a1112

*tools/base updated to 6073467d7d780083a6a61c0be783b55e5a81ed6a
Update r8 jar specification after update

We have updated the location of the api database

Test: //tools/base/gmaven:tests
Bug: n/a

Change-Id: I1a62d98de945b3eff213d40e97fdf3afb73f7087