Autoupdate of 2 submodules
*tools/vendor/google updated to 27567abca79d727d81023bc5620be18c97a79e7c
Move filter instantiation to cidr

OcOptionsApplicabilityFilter is implemented inside Cidr. If we
instantiate it from android-ndk, we get a cross-plugin classloader
error (when built with bazel).

This change moves it into cidr, to avoid such errors.

Bug: 226883449
Test: bazel build android_studio; run studio.sh; do search
Change-Id: Iac176fc8acfee4e4054665e626ab6b9d33d95251

*tools/vendor/intellij/cidr updated to 32f73e77845944a1ba4f7da20d75851c446b3b51
Move filter instantiation to cidr

OcOptionsApplicabilityFilter is implemented inside Cidr. If we
instantiate it from android-ndk, we get a cross-plugin classloader
error (when built with bazel).

This change moves it into cidr, to avoid such errors.

Bug: 226883449
Test: bazel build android_studio; run studio.sh; do search
Change-Id: Iec80f8877523ff5a55bdb6a2228844f68e4a3430