Autoupdate of 2 submodules
*tools/analytics-library updated to c68ec78e06b3942871ed21919d2ef678186d1b07
add new ScopedArtifacts enums.

some new methods were added that required enum values

Test: Existing
Bug: N/A
Change-Id: Ib11e99a8c4185adc29826086101901f64ddc792b

*tools/base updated to de5bcd86c3c3894a4c149e7417dbf915c5921972
Add support for getting all module classes

you could already get all module classes by using the
project_classes_dirs and project_classes_jars but that was
not very convenient especially if you needed to transform all
classes.

Instead, we are now proposing a new API dedicated to scoped
artifacts. A scope artifact can be scoped to either the current
project or the full dependency list. So far only CLASSES can
be scoped.

The API allows for getting/appending/transforming/replacing as
for non scoped artifacts.

When getting the current value of a scoped artifact, a Task
must provide two input fields, one for a list of jars and
the other for a list of directories. This is necessary as
inter-project publication is done via a FileCollection in
Gradle, we can get either of those on queries.

Test: added unit tests, API tests and integration tests
Bug: N/A
Change-Id: I8833da2b5a545ecdf91d440d6304762a5217ec53