Autoupdate of 2 submodules
*tools/adt/idea updated to dcfcccf7db19ec202e44157b416e309fc58d611b
[SDK Index] Use severity when reporting issues

There is a change in the SDK Index proto to now include a severity field
in the version labes. If used, this severity can be BLOCKING or
NON_BLOCKING and the meaning of it is wheter or not using this version
will block or not publishing. This new field defines the following types
of isses implicitely (sorted by severity):

1. Non complaint issues (always blocking, regardless of Severity field)
2. Critical issue (message from developer) with BLOCKING severity
3. Outdated library with BLOCKING severity
4. Outdated library with NON_BLOCKING severity (after some time it will
   be blocking)
5. Critical issue with NON_BLOCKING severity (will stay as NON_BLOCKING
   unless authors change it)

This change adds two new lint inspections to match changes in ag/18322382
and generates the PsIssues for PSD accordding to the severity in the
previous list.

Bug: 224643959
Test: Manual
Change-Id: I08afcdf43cdcec0fc6d69aa72c8733068d9fe501

*tools/base updated to de2b9f0332c6febd2967c5e72a40c0bb53cdd19c
[SDK Index] Use severity when reporting issues

There is a change in the SDK Index proto to now include a severity field
in the version labes. If used, this severity can be BLOCKING or
NON_BLOCKING and the meaning of it is wheter or not using this version
will block or not publishing. This new field defines the following types
of isses implicitely:

1. Non complaint issues (always blocking, regardless of Severity field)
2. Critical issue (message from developer) with BLOCKING severity
3. Outdated library with BLOCKING severity
4. Outdated library with NON_BLOCKING severity (after some time it will
   be blocking)
5. Critical issue with NON_BLOCKING severity (will stay as NON_BLOCKING
   unless authors change it)

This change adds new issues to reflect them.

Bug: 224643959
Test: GradleDetectorTest
Change-Id: I308f88a9ac0e9a22c4c0672d3a019d744a268cb8