Autoupdate of 2 submodules
*tools/adt/idea updated to 4c2f29cabc957316134812c188c96c17768ea0b4
Fix ApkFacet serialization

...and deserialization to/from disk by manually marking the
ApkFacetConfiguration as "modified", which causes the
configuration to be written to disk.

This is verbose, and error-proje, but it's better than the
feature being broken.

Bug: 178237380
Test: manual; save/load APK project; APK reload
Change-Id: I1fba21341bb62ce6dd26e1dc1fc6ac46a6a76fe5

*tools/vendor/google updated to e541675710344fb631db8b29e46c62a1250e1d83
Fix ApkFacet serialization

...and deserialization to/from disk by manually marking the
ApkFacetConfiguration as "modified", which causes the
configuration to be written to disk.

This is verbose, and error-proje, but it's better than the
feature being broken.

Bug: 178237380
Test: manual; save/load APK project; APK reload
Change-Id: I854778fad7013a5118c45a3e620ecb62a01b4e86