Remove over-strong run configuration validation check
It is legal for library projects to exist without an
AndroidManifest.xml file, as everything can now be specified in the
Gradle Dsl.  Remove the check for existence of AndroidManifest.xml in
the run configuration, and replace it with a check that the package
can be found from the ApplicationIdProvider.

Bug: 232420210
Test: added
Change-Id: Ib4711374454ee6ddafa93272c5bed78c4f382925