Autoupdate of 3 submodules
*tools/adt/idea updated to 3c84321a1c2d0abbbd0e83d220b2319f3ad7c62e
Add error code to ATTACH_ERROR logs

When we log that an error happened during an attempted connect we
store how far in the connection steps we got. However we do not
have an error code for what went wrong. Add this.

Fixes: 231745793
Test: Added unit tests
Change-Id: If87ae0c462abc48c6d4f22788dd9b4cc98eddfd9

*tools/analytics-library updated to 3710f4c14bbc2cb4fbe4d4794f141bc7ab40d12d
Update studio_stats.proto

Logs-Approval: cl/449256202
Bug: 231745793
Test: N/A
Change-Id: I291c149810ccf90791b5a72e9f493b6982827f6e

*tools/base updated to 48be27b1109a544dd7ae95a124f6dcb21167a16e
Add an error code to StartFetchResponse

This will allow Studio to include the error code in analytics.

Bug: 231745793
Test: Added unit tests
Change-Id: I75c9fcacb66484ea325f00a9e66e3ad88893612d