Autoupdate of 2 submodules
*tools/adt/idea updated to 09547018d9c0a4e7ddbbb04f7438a72d1dbaaf77
[Device Manager] Support AVD renaming

Instead of using the AVD Name as an ID
use the name of the folder used to
store the AVD info. The folder name
doesn't change when the AVD name changes.

Fixes: 232393311
Fixes: 225375165
Test: Updated Unit tests
Change-Id: I7b6db1bb0619141c111b2b1788ece8a529e7008c

*tools/base updated to 9f98571a9f6914fa6816f0c1f006350030adefcb
[Device Manager] Support AVD renaming

Instead of using the AVD Name as an ID
use the name of the folder used to
store the AVD info. The folder name
doesn't change when the AVD name changes.

Fixes: 232393311
Fixes: 225375165
Test: Updated Unit tests
Change-Id: Ie19536b9051f0cd634c70b0b93a45e0052c0bb1a