Autoupdate of 2 submodules
*tools/adt/idea updated to d0d7a4d283ad9ca7f50d7ee9b73614ee7a943b26
Refactor AvdManager to store ILogger as a field

Bug: n/a
Test: existing
Topic: rosej-avdmanager-cleanup
Change-Id: I2b57a9effdddc4d5f2a55fd0f369e0e067c22955

*tools/base updated to 984e9eca1c6c6e9fc6f603724cf2f05850a32be7
Refactor AvdManager to store ILogger as a field

All callers (AvdManagerConnection and AvdManagerCli) pass the same
ILogger instance in all cases. Thus, rather than passing this parameter
everywhere, store it as a field. This also allows simplifying the
DeviceManager field.

Store the cached valid AVDs as an ImmutableList, rather than an
array (which is vulnerable to mutation by the callers).

Bug: n/a
Test: existing
Change-Id: I3be045e81930f61afc37ebb020580a3c74383810