Autoupdate of 2 submodules
*tools/adt/idea updated to 61d9a70f42ce98e849c63708caac2b7beb4fadf5
Use new refactored string resource escaping utilities for clarity and efficiency. This will make updating the tests for these classes easier.

Bug: 232444069
Test: Currently covered by existing tests (the point of this project is to fix these tests)

Change-Id: I48ec0d34199e64cbf4812fc1b984203af5dc9d64

*tools/base updated to 9fa4dbf52361853bc8aaa577fe1a1305a504ec52
Refactor string resource library for clarity and efficiency.

This change allows these to all be singletons so that they need not be passed around all over the place when there is only really one instance.

Bug: 232444069
Test: Covered by new unit tests
Change-Id: Ib15cfb3d3879e157a4fabec74fe3ef20e8d5bf7e