Use single thread executor in the test gRPC service
Beofre the CL, the default executor for gRPC service was a cached thread
pool executor, which means gRPC messages sent by the test target were
processed on multiple threads out of order.

To preserve the message order, this CL specifies a bounded executor with
max_thread=1 when building the test gRPC service.

Bug: 233372677
Test: rely on the presubmit
Change-Id: Ie3e2da5bf7e8c5cffcfac731741c6df566f7c96b