Autoupdate of 2 submodules
*tools/adt/idea updated to 9942128eb6620adddd23c9ea5b186026cc1aa767
Do not include transitive tasks when building tests

For some outdated reasons we included tasks on main artifacts of all
dependency modules directly when building tests. Most of them do not
need to be included explicitly. Trust Gradle to build right modules
instead.

This change removes some but not all not needed tasks.

Bug: n/a
Test: snapshots updated showing unneeded tasks removed
Change-Id: Iee95224eb60bde3c231856cc85c5705c5cb9ca33

*tools/vendor/google updated to 767ba4845b8e86afe04e86b73520c576e009ecdd
Do not include transitive tasks when building tests

For some outdated reasons we included tasks on main artifacts of all
dependency modules directly when building tests. Most of them do not
need to be included explicitly. Trust Gradle to build right modules
instead.

This change removes some but not all not needed tasks.

Bug: n/a
Test: snapshots updated showing unneeded tasks removed
Change-Id: Ifda603c0ffc19e4949192c0c60de338b79f41ee7