Autoupdate of 3 submodules
*tools/adt/idea updated to 90c360a1c9793e47b07291922025b2be6a8ee7d6
Include application ID in the tooling model

Allowing us to simplify the implementation of getting the Application ID
in Android Studio, at the expense of not allowing it to be set as the
output of a task, which we can't think of a use case for anyway.

See I39b718b2b3b624371cc040ae1043f1209368cd51

Bug: 231946504
Test: SyncIssueUsageReporterTest, to follow
Change-Id: I271790778e868af61e3c494a86f29d81ca7b7cb4

*tools/analytics-library updated to b61ccbd2e814e3262ce19d88069d355c759bd73a
Include application ID in the tooling model

See I39b718b2b3b624371cc040ae1043f1209368cd51

Bug: 231946504
Test: Updated model snapshots
Change-Id: I88bb3e9f471709ea94f94869dae3a1a663dacb3c

*tools/base updated to 6c1066757775ef444cf18979afd19f1ea3627c2c
Include application ID in the tooling model

Allowing us to simplify the implementation of getting the Application ID
in Android Studio, at the expense of not allowing it to be set as the
output of a task, which we can't think of a use case for anyway.

Bug: 231946504
Test: Updated model snapshots
Change-Id: I39b718b2b3b624371cc040ae1043f1209368cd51