LE: Close channel on timeout
A command ran by the Installer can timeout on the device *after* the
Installer Client timeout on the Host machine. When this happens, the
installer command writes back an InstallerResponse which remains in
the OS socket buffer, dormant. This is what we call a "desync".

From there, a request get the response from the previous request
(and the current response go to the OS socket buffer). To break this
cycle, we close the channel on timeout which will trigger a clean
installer respawn on next request.

Test: AdbInstallerTest, InstallerTest
Fixes: 232146753

Change-Id: I794c81ca7db1c1b522143cb810c0ea8f81ea0193