Autoupdate of 2 submodules
*tools/adt/idea updated to 5d486b394a7cf5b21b44b976ad7cd189bbb41b7e
Implement test around updating

The test performs these steps:

- Unzips Android Studio
- Bypasses most of the setup process (e.g. the Welcome Wizard and SDK
  installation)
- Creates a patch
- Starts a file server to serve the patch to Android Studio
- Automates Android Studio to check for, download, and install the
  patch
- Ensures that the patch was successful

The test will likely be flaky at first as we figure out how to improve
both the framework and the reliability of the test itself.

Bug: 229769891
Test: these changes are to the test framework and tests
Change-Id: I99ceca59ee64512107d089411f44f667d3a3ece4

*tools/base updated to 9b29c2b6dae279e1c576afd64426e94b4b50116e
Implement test around updating

The changes in this repo are only to support the update test in
tools/adt/idea.

Bug: 229769891
Test: these changes are to test-utility code
Change-Id: I21f0b2747aa496285460bf17238e323608a6840a