Upgrade unselected variant config detection
Currently, we receive less information about unselected variants than the selected variant. Namely, we do not receive the list of flavor dimensions, or the build type.

This change extracts those from the variant name, under the assumption that the variant name is the flavor dimensions, followed by the build type, camelCased and concatenated.

Test: Covered by changes to test cases
Bug: b/233416315
Change-Id: I66c89deee2833d859fb7f16461a68b6493ef35c4