Autoupdate of 2 submodules
*tools/adt/idea updated to 228f9c079946b1e3013eb8cf4178adf920332562
Always get application ID from the model

The application ID is now directly supplied in the model for AGP 7.4+

For previous supported AGP versions, a new injected model builder can
fetch application IDs using the legacy variant API.

This also results in application not being avaliable for components that
don't actually exist - so if the test build type is debug, as it is by
default the test application ID will not be avaliable if release (or any
other variant with a different build type) is selected.

Injected model failures are reported sync issues so behavior is
consistent. While the new API was avaliable before AGP 7.2.0, Sync V1
always calls get on the value, so only AGP 7.2 & 7.3 currently allow the
application ID to be dynamic anyway.

This change does not attempt to clean everything up to remove the
dynamic application ID fetching codepaths, that will follow.

Bug: 231946504
Fixes: 184245551
Fixes: 184422212
Test: Updated snapshots
Change-Id: I17ed4d94aded3db658d47554d0d8b9f2cf03516e

*tools/vendor/google updated to 04e5363248839289dd275b6af02c563363687e79
Fix Firebase test now application id is avaliable

Previously the test was dependent on the fallback behavior
of picking the first firebase configuration if the application
ID was not avaliable.

Bug: 231946504
Test: FirebaseAppManagerTests.`allAvailableApps change upon gradle sync`
Change-Id: I821fc0c8bd06252e3f1a6d9278a9ece24312b57e