Replace "empty" ResourceItemEntry with null.
ResourceItemEntry hides some behavior by offering an alternate constructor
and a nullable field. It is simpler to have this be a straight data class
without any nullable fields or default behaviors (e.g. "" for string fields)
and instead let the layer above handle the case where there is no default.

Bug: 232444069
Test: Existing presubmits.
Change-Id: If35beb4e9f7a5585f095d72fdf0a4a24163be70a