Autoupdate of 2 submodules
*tools/adt/idea updated to 662487c76facac763ef36723c2e6e8f6535f3e79
Move StringsWriteUtils.createItem into DalUtils.java which is the only
place it is used.

This method is not used in the Translations Editor so clearly has no
business being in here. Instead, move it to the one place it is used
where it can become an implementation detail of an already-tested
method.

Remove the tests for this method as it is no longer part of
the public API of StringsWriteUtils.

Bug: 232444069
Test: Existing unit/presubmit tests.
Change-Id: Ic7658c21da92371291b0e57eaa2ff23a89338dce

*tools/vendor/google updated to 106bac344b5c2d7ab4ef70d8cf7a31e8082b1a05
Rehome method from StringsWriteUtils that is only used here.

Bug: 232444069
Test: Existing unit tests.
Change-Id: I48c920471e09fa2c919e978c772f5303b49b8350