Autoupdate of 2 submodules
*tools/adt/idea updated to 6a9dfff171e974e291097a0332f15be9746cba63
Create StringResourceWriter to replace StringWriteUtils.

This will give us a mockable interface we can inject into code that uses
it so that we can test without needing to worry about using the file
system or testdata or whatnot.

Also add tests for lots of parts of this class that were not
previously tested.

Bug: 232444069
Test: StringResourceWriterTest.kt
Change-Id: Ia64ab5907c689c1d1a1754f866e0ccc6ee61114f

*tools/vendor/google updated to 44f322837fcd80ac1b7c27136ed38a11eb90d6d1
Use new StringResourceWriter instead of old StringsWriteUtils.

Test: Existing unit tests
Bug: 232444069
Change-Id: I14ad43be8edb0feb59aaec2b6637ace7679ed938