Autoupdate of 2 submodules
*tools/adt/idea updated to 89d2c6abef2f82140e67a320d8ca3d9f1ba560ad
Update generated files for permissions inspections

Manually update message text for RESERVED_SYSTEM_PERMISSION and
SYSTEM_PERMISSION_TYPO.

Bug: 232652234
Test: N/A
Change-Id: Ie7af80bfc6156494d6fed550f6b5d32e6865b0ac

*tools/base updated to d7d363e38523283ea8780eeefa8947b1ef8bc81f
Update lints for reserved system/sdk permissions

There are two types of "reserved" permissions:
1. System Permissions (hard coded in `SystemPermissionDetector.java`)
2. "Platform" or "SDK" Permissions (from an SDK's
   `android.Manifest.permission` class)

Those two lists overlap with one another.

Neither type can be redefined as a custom permission by an app
developer via a `<permission>` declaration.

Both types could be subject to typos in `<uses-permission>` declarations
or components restricting access via `android:permission`.

One might refer to the union of these two lists as "Standard"
permissions, insofar as they are the opposite of custom permissions.

With this change, we update the prose of the related issues to reflect
the fact that the suggested values may come from two places. We chose
not to update the issue IDs since the checks have already been released,
and the IDs could be used in suppressions.

- reduce MAX_EDIT_DISTANCE from 4 to 3 to minimize any false positives
- Rename some helpers/variables accordingly
- Pull some of the typo logic into a named method for readability
- Add a negative test to prove there will not be false positives on
  custom permissions that happen to share the same suffix as a platform
  permission.

Bug: 232652234
Test: PermissionErrorDetectorTest
Change-Id: Id5b3e71ed393cacb201a7628774684191b4c7fab