Bazel Kotlin worker: clear jar cache after jobs
See the bug and added code comment for details.

If we wanted to be extra cautious and preclude this class of bugs,
then we could use the --worker_quit_after_build Bazel flag
(just like we already do on CI). However, this flag has a performance
penalty for repeated builds.

Fixes: 234840104
Test: manual (using discovered repro scenario)
Change-Id: I2f394479ed711ce776202638aec6ac69f23e48e7