Changes patch test to use version numbers
Instead of modifying a properties file, this changes
the test to use the same mechanism to update the version
number to change the target studio. This way we can use the version
API to assert that the build has changed.

This change integrates the fromDir change which needed some tweaks:
 - removed incorrect use of TemporaryFolder and made model classes
   not depend on junit classes.
 - added a reset method to the std streams so when we run a new
   process we can close the open streams
 - fixed AndroidStudio and PatchMachinery to distinguish between
   workDirs and studioDirs

Note the std streams are still incorrect as they are left open

Test: this is for tests
Bug: NA
Change-Id: I5bac88e1b9c81a1f34cd7d95503b760ede2cbfe7