Autoupdate of 2 submodules
*tools/adt/idea updated to f4a1917254e986cc730505885439603ab9679500
Add support for multiple devices

Enable ForegroundProcessDetection to handle multiple devices.
Each device corresponds to a stream.

Each time a device is connected we register an event listener
for LAYOUT_INSPECTOR_FOREGROUND_PROCESS events from that device.

When a device becomes the selected device, we send a
START_TRACKING_FOREGROUND_PROCESS event to it. If there was a
selected device before, we also send a
STOP_TRACKING_FOREGROUND_PROCESS to that device. This makes sure
that the on-device polling happens only on the selected device.

The SelectedDeviceModel introduced in this change will become
more useful later on when for the UI we will need to know at
any time which device we are listening events from.

Test: ForegroundProcessDetectionTest
Bug: 232094819
Change-Id: I73483f19b33c67b52ca6bf8cf5ec05f2cd5cf085

*tools/base updated to 61a4090ff96e23c90504462bfa16f284ff6d118a
Initialize Layout Inspector ForegroundProcessTracker

Initialize LayoutInspector ForegroundProcessTracker
in the transport.

Test: dynamic-layout-inspector/foreground-process-detection/test/ForegroundProcessTrackerTest.java
Bug: 232067042
Change-Id: I5130a1bba8e95a16619d791d8f5d7c9ae729e560