Ensure event name is not omitted in end-events
event name should be omitted with `TRACE_EVENT_BEGIN` macros
but not with `TRACE_EVENT_END`. This is taken care already at the
public macro level. However it still makes sense to ensure trace
implemention doesn't omit event-names in end-events. Which was also
supported for StaticString event-names. However, this was not taken
care if event name is DynamicString (aosp/2111624)

Change-Id: I9cb6bc157f3bae47115eefe178d6f366d9826b18