Relayout the Compose Preview list only once before rendering
Every time a new model was added to the Compose Preview DesignSurface,
a relayout was being done, but not when an existing model was being
reused. This caused for some Previews to "jump around" the panel
before starting to render the updated models.

This change modifies the DesignSurface so that the update on the
SceneViews is not always done after a new model is added, so that the
Compose Preview panel is updated only once before starting the render
process, after adding/reusing all the corresponding models.

Bug: 232718495
Test: manually verified
Change-Id: I236581ce92e70246fe206e9df237d55f388ae06f