Prevent multiple event round trips upon issue selection.
Whenever the selected issue changes externally(i.e. by clicking on the
table view row) we get multiple events firing which causes 2 unnecessary
round trips to the model and back to the table view.

To avoid this, the change disables change listener on the selection
model since it is redundant.

bug: none
Test: none
Change-Id: I245fbe2b3d5ef27980b89f8a954d7123035860b2