Autoupdate of 2 submodules
*prebuilts/r8 updated to c4c9f567b1bda961dfcdcf42b5fa3887a93b8f44
Update D8 R8 to 4.0.10-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/4.0.10-dev
Test: ./gradlew check
Bug:

Change-Id: I0d3e3786b1b1b5aca36963b888df7faa26b2ae37

*tools/base updated to 88997a2a257e7c270465b4ba0317bbaff782d2aa
Update DynamicAppTest and SwitchMultidexTest expectations

Test: //tools/base/build-system/integration-test/application:coupled_tests
Bug:
Change-Id: Ic4ea9a68fb83113fd6db4e801c9aff1903109b22