Autoupdate of 2 submodules
*tools/adt/idea updated to 104aad63723438d1445c664e1c82d209aca15998
Execute and handle handshake

This change adds logic that initiates the handshake with the on-device
library each time a new device is connected.

If the handshake is successful the device supports foreground process
detection. Only in this case Layout Inspector starts foreground process
detection.

If the handshake is not successful, we fall back to the old
implementation of Layout Inspector, where the user has to hand-pick the
process they want to inspect.

Test: ForegroundProcessDetectionInitializerTest, ForegroundProcessDetectionTest, SelectedDeviceActionTest
Bug: 232067042
Change-Id: I642c2b1e0cc55436cc2609c932dcd9883b8e948c

*tools/base updated to e61e5c6aa30fe86d077118f21d86fe85a579ea4e
Add handshake logic

This change introduces logic to handle hanshake commands.

Before Studio starts asking for foreground process information
it needs to make sure that foreground process detection is supported.

For this reason we need an handshake where Studio asks the library
if the device supports foreground process detection.

Foreground process detection might not be supported in cases where
dumpsys is not avilable, its output has been changed or if for any
other reason the library fails to detect the foreground process.

Other than making Layout Inspector more resilient to bugs in the
library, this change is mostly useful to ensure backward
compatibility. Since forward compatibility will be ensured by a
CTS test.

Test: ForegroundProcessTrackerTest.java,
foreground_process_tracker_text.cc
Bug: 232067042

Change-Id: I1c53baee44a3a6c23485d3e57b271dba7b149b82