Autoupdate of 2 submodules
*tools/adt/idea updated to 766d7e43f0c28f17483ad4ab07ef699fc91d795d
[SDK Index] Make lint and PSD messages consistent

... by using the message texts from GooglePlaySdkIndex instead of
creating them in PsAnalyzerDaemon.

Different messages were introduced when using the severity from
the proto, but the changes were not consistent between PSD and
lint. This change causes the messages in PSD to be generated from
the same methods as the messages in lint.

Bug: 224643959
Test: GradleDetectorTest
Change-Id: Iafe082aff12fb7947e8c16d26bb37c1564547019

*tools/base updated to cdf47054b8424f975394274ff7d2dd8bc4824473
[SDK Index] Make messages consistent

... between lint and PSD.

There are some slight differences between messages shown by lint and
PSD, the differences were introduced when the severity from the index
was used.

This change makes text from those issues to be the same in lint and PSD
by generating those message in  a single place (GooglePlaySdkIndex).

Bug: 224643959
Test: GooglePlaySdkIndexTest
Test: GradleDetectorTest
Change-Id: I28eefa8017cd7bef311f089a214e7a001709c05f