Autoupdate of 2 submodules
*tools/adt/idea updated to e0365f29e33d134c0aa0b8c5a8117340efa80a99
Send addRule command upon user action

This change makes it so that we immediately send an addRule command to
the inspector when user clicks the add button in the UI.

Secondly, changed usages of addRule command to updateRule when it's
appropriate. (they are treated the same by the agent, but we should
still call the right command)

Thirdly, this change enables and disables rules by sending an update
evente, as opposed to adding/removing rules.

In addition, this change introduces logic to differentiate table events,
so we don't send reorder command when the action performed isn't related
to reordering.

Test: RuleDetailsViewTest
Bug: n/a
Change-Id: If46dc7ba08c00e660afd7802bdc532fd008987b9

*tools/base updated to eb2e10284998b375ad1bf772a8e1dcb587c046d8
Support disabling of rules

Previously we disabled/enabled rules by adding/removing them from the
service, however that made the reordering logic more complicated.

This change introduces a boolean field in rules. The interception
service filters out rules with the field set.

Test: new InterceptionRuleServiceTest
Bug: n/a
Change-Id: I27faabe06508a07843787b4e2ab700539fa91723