[TemplateTest] Speed up test
Test was creating a new adhoc TestFixture, configuring it and
then clean it up. This was used when the Template was tested
multiple times, with different parameters (like different API
levels), and was needed to create multiple projects and build
them separately. For performance reason the templates are now
only tested once and this adhoc setup can be removed.
If in the future we need to test the same template with
multiple setups, a solution will be to create multiple
modules and then build all in one go (this should keep the
build time of the project from growing too much).

Bug: 234749894
Test: TemplateTest
Change-Id: Ib8bf6610ea79a99089739d742379bc31c4b75027