Autoupdate of 2 submodules
*tools/analytics-library updated to 19ea8a067f3104a5b7504e2b432995ffbe2994bf
Add task to build the privacy sandbox SDK APKs

Bug: 235469089
Test: AnalyticsUtilTest. Use to follow
Change-Id: If1a216dc822256fe6e32a74b1e557db465090587

*tools/base updated to 84adfbbdc107ebcf98db6371ca96a21d7480e527
Privacy sandbox sdk: Generate APKs for local dev

Fix the incorrect publication of the runtime classpath in the fused
library base plugin to not publish dependencies between projects too, as
well as adding support for publishing the ASB between local projects (as
a placeholder for the asar).

Adds a task to trigger the transformation of the privacy sandbox sdk in
to APKs and extract them to an intermediate directory.

Bug: 235469089
Test: Covered by PrivacySandboxSdkTest
Change-Id: Ica169f3898d91c12e945926f044a1c85ae06dd7d