Update D8 R8 to 4.0.12-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/4.0.12-dev
Test: ./gradlew check
Bug:

Change-Id: I7d2df352bf250960a6b7b36816e0482e66db6154