Allow Logcat To Be Initialized By Game Tools
In order for the logcat to be initialized by game tools, two things
were needed. One was access to the logcat window class. Making the
class public caused too many issues with other classes needing to be
public, so a factory class was created instead.

Second, the project was not getting set correctly on events, meaning
that the screenshot action would fail due to a missing project.
However, the project is available when the UI element is added.

Bug: 228092950
Change-Id: I3234ff8949b03790573daee64a6f489123d26744