Register Image File Editor For Game Tools
In order to view screen captures we need to register a file editor for
PNG images. Normally we would just depend on the images plugin, but we
have disabled plugins for game tools.

Instead, we register the editor and the necessary services needed to
view the screen captures (PNG images).

Bug: 228092950
Change-Id: Ib4a5f226affbb402754a822a40346ae15dd9247e