Add check for specific JDWP error at startup
When attempting to debug Android Studio locally, it's possible to hit an
error where JDWP can't bind to an address because it's already in use.
This information currently only prints to Android Studio's stderr, which
makes it hard to find as a developer. Even after discovering it, it's
not obvious what to do about it without reading documentation.

This change diagnoses the issue automatically and provides remediation
steps.

Bug: N/A
Test: this is for the test framework
Change-Id: Ic3940084cab2390b7d46659a8015e5826041b30a