[DM] Override device index instead of using parameter
Previously, we were passing the device model column index into the
Virtual and Physical device table superconstructors, and then assigning
that inside the DeviceTable constructor. This causes an issue because
the JTable initialization calls createDefaultTableHeader, which in our
Virtual/PhysicalDeviceTable implementations, calls
deviceViewColumnIndex(). All this happens before the rest of the
constructor body runs, so myDeviceModelColumnIndex was not assigned yet.
This hasn't been a problem yet because our device index was always 0,
the same as the uninitialized value, but that might change soon.

Bug: None
Test: None
Change-Id: Ibfaba0cce62df5347d76caad04f0242dfb14e373