Not found threading notification group in tests
For unit tests, running a threading agent and not including 'threading-checker.xml' in the test's plugin.xml could result in NullPointerException since a "Threading Violation Notification" group wouldn't be found.
It's ok to skip surfacing threading violation notifications in the tests since we'll rely on a different mechanism for verifying threading correctness in tests.

Bug: 228841408
Test: Unit tests
Change-Id: I3a01e66be103ee0561a2d329937e38e21ceaa8bf