Autoupdate of 2 submodules
*tools/adt/idea updated to 625ad3652b16db254c5a3aa36f8a40b7e83a3d6c
Remove default language to region mapping usages

The locale manager had a lot of heuristics to try to guess the
most appropriate region to associate with a language.

In some cases this is simple and straightforward -- a language
is spoken in exactly one region. But it's often more tricky
than that, and we had heuristics which tried to pick the best
match based on matching things like the current timezone of
the user with the region timezones etc.

The primary and original use case for this was to show flag
icons next to resource files. For example, if you have a number
of strings.xml files, we might show a French flag next to the
values-fr/strings.xml file and so on.

However, trying to associate regions with languages is just
too tricky and can sometimes upset users, so a long time back
we stopped actually showing the flags by default and hid it
behind an option, and at some point the option disappeared too.

There were a handful of remaining minor usages, primarily to
help sorting, but these aren't necessary, so this CL deletes
the remainder of this code.

Test: Existing tests (some updated)
Bug: Not filed
Change-Id: Ibbd2c2cdfd2e612dc8e0e8f5253d4a31d6fedd84

*tools/base updated to 81632f8483e8ad7861f41ef301e36862c893a953
Remove default language to region mapping and usages

The locale manager had a lot of heuristics to try to guess the
most appropriate region to associate with a language.

In some cases this is simple and straightforward -- a language
is spoken in exactly one region. But it's often more tricky
than that, and we had heuristics which tried to pick the best
match based on matching things like the current timezone of
the user with the region timezones etc.

The primary and original use case for this was to show flag
icons next to resource files. For example, if you have a number
of strings.xml files, we might show a French flag next to the
values-fr/strings.xml file and so on.

However, trying to associate regions with languages is just
too tricky and can sometimes upset users, so a long time back
we stopped actually showing the flags by default and hid it
behind an option, and at some point the option disappeared too.

There were a handful of remaining minor usages, primarily to
help sorting, but these aren't necessary, so this CL deletes
the remainder of this code.

Test: Existing tests (some updated)
Bug: Not filed
Change-Id: If01e58ef42cf53c538f2a55f420378ccb7ba225e