Refactor test ModuleInsightsController into a rule
This change extracts out the setup and teardown of test
AppInsightsModuleController so it can be reused in other tests.

To avoid race between consumption of state and actions taken by the
tests, all states generated by the controller are now put in a Channel
to be consumed by tests. Tests are now expected to consume all states
generated by their actions. This fixes the flakes that have been
happening.

Test: updated existing tests.
Bug: 234923248
Fixes: 235625101
Fixes: 235888866
Change-Id: I7611698f1febfd8cfd9be994c7ef12d664b96cb0