Fix UX when PreviewParameterProvider throws an exception
When processing a Compose Preview that takes a PreviewParameterProvider,
we were logging eventual exceptions as debug and returning an empty
sequence in ComposePreviewElementTemplate#instances(). As a result, if
an exception was thrown in the PreviewParameterProvider constructor, the
Compose Preview panel would be empty and users would be clueless, since
the exception was also not logged.

This issue is fixed by returning a SingleComposePreviewElementInstance
that fails to render in ComposePreviewElementTemplate#instances(). The
visual feedback will be the render errors panel, so users will know the
preview failed to render. In ComposeRenderErrorContributor, we now check
if the render error is caused by an exception thrown in the constructor
of PreviewParameterProvider and handle it accordingly, displaying an
intuitive message to the user.

Furthermore, we now log the exception as a warning, so users can check
the IDE logs when they're not sure what's wrong with their
PreviewParameterProvider implementation.

Bug: 234054935
Bug: 235848118
Test: ParametrizedPreviewTest, ComposeRenderErrorContributorTest
Change-Id: I088cbc6cdb69215e812ff38c36f9a42d700fcf19