Autoupdate of 2 submodules
*tools/adt/idea updated to 84020f57b3b6e8fca4cdc06ef5ec20f43983f227
Introduces AndroidSystem to the test framework

This class represents the system the user is running
studio on. It contains the file system, the display,
the environment variables, etc.

It provides an abstraction of the common operations
that are done on the system to ease the writing of tests.

Note that complicated tests, like UpdateTest, still
can use the raw apis for more complex flows.

Test: these
Bug: NA
Change-Id: I5c987ee26af2d9feb877c66e7e701de29b564424

*tools/vendor/google updated to 0b93aabb7f6d9bd0f9aef8fb609ca07087296aff
Introduces AndroidSystem to the test framework

This class represent the system the user is running
studio on. It contains the file system, the display,
the environment variables, etc.

It provides an abstraction of the common operations
that are done on the system to ease the writing of tests.

Note that complicated tests, like UpdaterTest, still
can use the raw apis for more complex flows.

Test: these
Bug: NA
Change-Id: I32908d29cadf50e820ab9cb43119104e2b738567