[Dsl,.toml] Adapt to Project Structure use of DslText
The Project Structure layer is not quite strict about separating
internal names for things from external names, and will call into the
Dsl model with strings of uncertain syntax.  Adapt to this by making
the TomlDslNameConverter idempotent (no prepending of "version." any
more) and making the Dsl element layer more aware of the semantics of
version catalog files, introducing special cases for the resolution of
references within version catalog files (the syntax requires no
prepended "version." for version.ref lookups) and then for translating
from the hierarchical notation to the string notation required in
Version Catalog files.

This all requires a more careful handling of scope of references,
particularly of references from models which have transforms: we need
to perform reference resolution with respect to the transformed
element rather than the raw element.  Introduce a getHolder()
interface function and provide appropriate implementations.

Finally, consistency is only maintained in the Project Structure
dialog if isReference() is accurate; implement it in our special
GradleDslVersionLiteral class.  (It is possible that the parent class
is also wrong in this respect, as nothing ever seems to call
setReference().)

Bug: 200280395
Test: added, manual test of the PSD
Change-Id: I9d5203e6df8aa9175502f9bda7cee9b70ec9931c