Fix Connection Assistant getState error
Change the exception to a warning. Since ag/17129435 made init run off
the UI thread, the state manager isn't initialized yet when getState is
called from the UI thread. Because updateButtonState gets called after
the state manager is initialized, we can just return the default
IN_PROGRESS state when it's not ready.

Bug: 235570310
Test: None
Change-Id: I7f52cf42527ffd2ead7799378cbc9df9dddca543