Autoupdate of 2 submodules
*tools/adt/idea updated to 296f8c89a73e668b79779ee617a375f2e906b661
Improve handhsake

If a device is connected to Studio while locked, the current handshake
implementation does not work. As the top activity wouldn't be found.
Therefore the handshake would terminate saying that the device does
not support foreground process detection, which might be a false
negative.

To work around this limitation, the handshake has been changed.
Now the handshake has three possible result states:

1. SUPPORTED
2. NOT_SUPPORTED
3. UNKNOWNW

This change handles these new results of the handshake process.

SUPPORTED and NOT_SUPPORTED are self explicative.
UNKNOWN means that at the moment we can't determine if foreground
process detection is supported or not.

When we get an UNKNOWN resposne, we should repeate the handshake
periodically, until it results in SUPPORTED or NOT_SUPPORTED.

Test: ForegroundProcessDetectionTest.kt
Bug: 232067042
Change-Id: If4979d94f9e61edc0b4cdfce093d8ee00f67add2

*tools/base updated to 97db61e2e37a2659f9caf8ab85cf12409da0f9cb
Improve handshake

If a device is connected to Studio while locked, the current handshake
implementation does not work.
`dumpsys activity processes | grep top-activity` would be empty.
Therefore the handshake would terminate saying that the device does
not support foreground process detection, which might be a false
negative.

To work around this limitation, this change enhances the handshake
protocol in this way:
If top-activity is not found, check if there are awake
(dumpsys activity activities | grep isSleeping=false) or sleeping
(dumpsys activity activities | grep isSleeping=true) activities.

If there are sleeping activities, but no awake activities,
it means that top-activity might be absent because the
device is locked. In this case the handshake terminates
with an UNKNOWN supported state.

If there are awake activities, it means that for some reason
the top-activity is not returned. In this case the handshake
terminates with a NOT_SUPPORTED state.

If there are no awake and no sleeping activities, it means
that dumpsys is not working as we expect. In this case the
handshake terminates with a NOT_SUPPORTED state.

Test: foreground_process_tracker_test.cc, ForegroundProcessTrackerTest.java
Bug: 232067042
Change-Id: I51ffbefeca54039da8064d49212ea44285e860f1