[AUA] Change the layout for UpgradeSyncFailed
Use minimal horizontal extent for the details so that we can have a
second panel to the right for the Sync error message, now in a
scrollable text area.  This ensures that the revert button is always
visible, no matter how large the sync error message.

Also rearrange the order of button and 'Local History' pointer as
suggested by UX, for all instances of the revert button, and add
spaces and remove trailing periods.

Fixes: 238470905
Test: existing, manual
Change-Id: If98c6f0b6f038e93e4871ea62739c2f47d22c2e1