Autoupdate of 2 submodules
*tools/adt/idea updated to 1c4e55dc193652410b56e38a0142eee32972f291
239554951: Update lint severity ordering

Lint changed the ordering of the enum values in Severity such that
they're in increasing severity order; this required updating
comparison code on the IDE side.

Test: IssueDataTest
Fixes: 239554951
Change-Id: I48559f63c70def3ed68babc2db170b77b9d21731

*tools/base updated to 2d9e34bc1e9ac4d835f8f783938cda6bf4fa6035
239554951: Lint's ignoreWarnings flag (-w) ignores warnings but not informational severity

In addition to fixing the bug, this CL reverses the order of the
severity enum constants such that their comparison order matches
the meaning of severity (e.g. error is more severe than warning).

Fixes: 239554951
Test: Unit test included
Change-Id: I9036a21cf1a6d66ab473631840eae6378d5490dd