Autoupdate of 2 submodules
*tools/adt/idea updated to 2663df0862f0c9416be9787319aca3cac2ba5d7f
adblib: update code for API change

Bug: 188067646
Test: Existing
Change-Id: I3d51a5895b2db61358a2e0bd87b94fefba2e5a92

*tools/base updated to 4e1d128d463cb3dd1faa19f0fba93fea86205d9f
adblib: cleanup usages of `shellCommand`

Remove shell extension functions that are superseded by the
`shellCommand` builder, which is more flexible. Also, rename
classes/functions related to the shell,v2 protocol to "shell",
and the classes/functions related to the legacy shell protocol
to "LegacyShell".

Bug: 188067646
Test: Existing
Change-Id: Idae6a9496ce0e555a9bdda9d248915000a0976a5