Fix scrolling in stacktrace panel
The rendering of the stacktrace in the stacktrace panel causes the caret to advance to the end. The fix to move it back to the start and reset the scroll position.

Also disables the showing of the vertical scrollbar inside the stacktrace editor to avoid situations where we show two scrollbars.

Test: manual
Fixes: 239756652
Fixes: 235361809
Change-Id: Ic45ead6eb5cf5c1d698e20ec4e62a622621b61d6