Log warnings insteads of errors
...to avoid triggering crash reporting.

For logging warnings in fetchOAuth2Token,
this is safe as those exceptions get
rethrown from the login plugin, and in AQI,
we catch these and print out the error message
in the "Issues" panel, then users can take
actions to resolve the issues.

For the rest not really on the path we make
rpc calls from AQI, warnings are still more
approriate than errors.

Bug: 241115746
Test: manual
Change-Id: I255492fed57a6944929c7c9e36ca14c2514b31da