Fix Android Studio sync for dependant Prefab modules
There are two changes here:
1) Modules no longer need to be C/C++ configured in dependency
   order.
2) Module-to-module references now cause the live version of
   header files to be accessed rather than the Prefab-packaged
   version.

To accomplish (1), this CL introduces the concept of a "header-
only" version of a Prefab package. The header-only version does
not require C/C++ configuration to execute and it is generated
at Gradle configuration time (which completes entirely before
sync).

Header-only prefab packages are sufficient for language-service
purposes because the IDE does not care about linker flags. It
doesn't actually know them.

Bug: 221231432
Test: Added test to cover the scenario
      Manually verified it works with AndroidX
Change-Id: I8740e5cba4d9987f80b6929955c01bdf2051e658