Fix Abseil perfgate regression
The regression was due to ag/19304352 which caused an extra
C/C++ soft configure. The reason was that some of the files being
monitored for triggering a C/C++ configure were in the build/
folder which is deleted on clean.

The fix is just to not monitor those files. None of them merit
moving them out of build/ into .cxx/ folder.

Bug: 240718091
Test: Is test
Change-Id: If533d04fe4a550c191fdc128c0ce2a591d97cc63