Autoupdate of 4 submodules
*tools/data-binding updated to 5da7487dca53d3fa25b41351de93d77a375a6610
Remove old Compose dependencies

Test: bazel test //tools/adt/idea/...
Bug: N/A
Change-Id: Ieb804ac99f38c8399799c6c40ce8de2546df4c07

*prebuilts/tools updated to 8d800494c6773a06484daacb0ea0f536feaf1e69
Remove old Compose dependencies

Changes related to the deletion of Comopose 1.1.1 and 1.2.0-alpha05. As
part of this change, androidx.test.ext:junit:1.1.0 and 1.1.1
are not needed anymore and removed as well.

Test: bazel test //tools/adt/idea/...
Bug: N/A
Change-Id: I5a26a5401f8baf0b7e45a2cf366ddfd065bfc872

*tools/adt/idea updated to 8604c0b9a367241444a3008f3b3d7a8a7bae66da
Remove old Compose dependencies

Changes related to the deletion of Comopose 1.1.1 and 1.2.0-alpha05. As
part of this change, androidx.test.ext:junit:1.1.0 and 1.1.1
are not needed anymore and removed as well.

Test: bazel test //tools/adt/idea/...
Bug: N/A
Change-Id: I9e96e0fc6d37aed1777e3d2a2288998b23ae840f

*tools/base updated to a7c609c9a75df5c8754c22b4d6239b6ced369344
Remove old Compose dependencies

Changes related to the deletion of Comopose 1.1.1 and 1.2.0-alpha05. As
part of this change, androidx.test.ext:junit:1.1.0 and 1.1.1
are not needed anymore and removed as well.

Test: bazel test //tools/adt/idea/...
Bug: N/A
Change-Id: I00b2a6e00af455e0000525c58a9c6c60af235f01