Autoupdate of 2 submodules
*tools/adt/idea updated to d0078c33493ecee4dad7c856778960cd155d4991
Add feature flag to disable keyboard event

... in profiler's Event Timeline

The logic takes place in the app process. For pre-O,
the flag is passed as a system property to control compile
time instrumentation. For O+, the flag is passed as a
field in the JVMTI agent's config file.

Bug: 211154220
Test: manual as this change involes many parts of Studio
      including build, deploy, running and interacting with
      the app, and running profiler and the daemon
Change-Id: I8beb55ea1e8d6fbcf15e49e4b97f0c70e968658f

*tools/base updated to 2b15f53e53ca18762eaed08d8ede1e4494797466
Stop supporting keyboard events in Event Timeline

For pre-O (API 25-), the instrumentation is done at compile
time. For O+, the instrumentation is done through JVMTI
at the beginning of runtime. Both approaches are updated.

We leave the door open to possibly enable the functionality
through a server-side flag for L - Q (API 21 - 30), in case
user feedback indicates a strong desire of it. If it
happens, we need to figure out a proper way to support it
for API 31+ as the implementation is out of sync of
platform.

Bug: 211154220
Test: Update existing tests as if the functionality is always enabled.
      Manually tested the flag because this change involes many
      parts of Studio including build, deploy, running and
      interacting with the app, and running profiler and the daemon.
Change-Id: Ib037bf98f65747fee8c6a32e18eeb2b4501a012b