Autoupdate of 2 submodules
*prebuilts/r8 updated to bc970493d67ceb963b8579535ed969ad6bec351c
Update D8 R8 to 4.0.21-dev

This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/4.0.21-dev
Test: ./gradlew check
Bug:

Change-Id: I3ad4a8f15eb742533b7be9664db66d98ee57c636

*tools/base updated to 8df05d19565afd025f1f230af3bbaa27b55eeccd
Upgrade UnsupportedFeatureDiagnostic warning to error

Test: coupled_tests
Bug:

Change-Id: Iad7edb6ca030d0d4d42a9693ec98854f8309299e