Implement ToolWindow Replacement For GameTools
When an event is fired from, the UI hierarchy is traversed (bottom to
top), checking if each element is a DataProvider. If an element is a
DataProvider, it is asked to resolve a data id. Normally the
ToolWindow resolves the "project" data id, but since we don't use a
ToolWindow, we need to implement a new top-level DataProvider.

This change was needed as our original change to LogcatMainPanel was
reverted in another change.

Bug: 241564903
Change-Id: I2bd20813f0239071ec5d8736dec286cacd048466