Narrow scope of tests disabled for b/240444529
The prior disabling of the test disabled the entire test.
This CL reenables the tests that weren't failing.

Result of running the test 1000 times after this CL is
applied:
https://fusion2.corp.google.com/invocations/5078ecae-eda2-44d7-ad4b-f7f2b71aee30

The two tests that are disabled are for the V1 sync model
which is deprecated and I think we can remove it when we
have the resources to do that work. The impacted users will
be those using modern AS with AGP less than 7.0.0.

Bug: 240444529
Test: Is test
Change-Id: I8a16664715835ceb69404a33e3496f20e66c51f7