[AUA] Bump required Kotlin plugin version
In AGP 7.3 and 7.4, while strictly speaking the required Kotlin
version is 1.5.21, if the `kotlin-android-extensions` plugin is
applied anywhere AGP will throw an error if the Kotlin version is less
than 1.6.21.  Rather than try to distinguish between the cases where
the extensions plugin is applied or not, unconditionally require
Kotlin 1.6.21 in the Upgrade Assistant; this will cause some projects
to be upgraded to a newer Kotlin than is strictly necessary, with the
benefit being that projects using the extensions have a smoother
upgrade.

Bug: 215545075
Test: added, existing tests adjusted
Change-Id: I15b013e63c49441778457bc1f365412ba3a70217