Update D8 R8 to 4.0.22-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/4.0.22-dev
Test: ./gradlew check
Bug: b/241347641

Change-Id: I6d2aad47957ed3235c1de826f928c275e361ade1