Re-enable test: CurrentAgpV2#testNewSyncKotlinTest
The flakiness of this test is linked to the following issue
IDEA-297059 introduced recently on the platform, given to the
intellij-community commit: dff59a2d1944e020dcb4e4d2c176c9b9a3b79721

The CommonGradleProjectResolverExtension.getGradleOutputDir() method
is picking the first source directory based on HashSet with the
addition that by removing the kotlin source this will ignore the
kotlin directory given that this one doesn't exist, discarding in
this way relay on the HashSet order (root of the issue)

Comparing the SrcDirs we have the following:
- Before
  - Available SrcDirs: [..main/kotlin, ..main/java]
  - Available SrcDirs Existing Files: [..main/kotlin, ..main/java]
  - Returned gradleOutputDir: Based on random ordering of SrcDirs
- After
  - Available SrcDirs: [..main/kotlin, ..main/java]
  - Available SrcDirs Existing Files: [..main/java]
  - Returned gradleOutputDir: Always java

This will guaranty that the test will always get the java source.
Similar to "kotlinKapt" test where even that this also apply
"kotlin" plugin the HashSet order doesn't affect it by the fact
that kotlin sources directory don't exist.

NEXT STEP:
We are going to create a patch on the platform and once everything
it's merged we are going to move kt.java file to its original path

Bug: 237389151
Test: Re-enable testNewSyncKotlinTest
Change-Id: I5d33106d32b58bf5746136981ab2d6f67333c549