Remove broken support for module level libraries
for local jars.

Since `moduleFileDirectoryPath` is not a path to the root of a Gradle
project we never create module level libraries just because they
represent a `.jar` file under a Gradle project root.

This behavior can easily be fixed, however temporarily removing this
code simplifies refactoring needed in order to setup IDE libraries
for gradle libraries present on runtime classpath only.

Bug: 243008075
Bug: 169916204
Test: android/testData/snapshots/syncedProjects/two_jars_V2.txt
Change-Id: Iee79ea004540a6cc6ff02aeedeb5aa3068573932