Remove confusing HalManifest::getMetaVersion()
The function intentionally always return the metaversion
of the current libvintf. It did not returned the original
version attribute specified in the XML file.

This ensured that XMLs produced by libvintf always bear
the latest meta version.

However, the function name is misleading in that it
may be interpreted as the original version attribute specified in
the XML file.

To avoid confusion, just delete it and write the constant
explicitly when generating XMLs.

Test: TH

Change-Id: Ib70b090254a94280827e7cb93d1655b771e6c086