Changed Mem. Prof. to use value to determine unit
The memory profiler tooltip values were losing precision when
using the range to determine the units to be used.
Now to determine the unit converted to/shown, it uses the value
itself, displaying the smallest unit its value passes the
conversion threshold of. For example, 0.5 GB will now show as 512 MB.

I also added that the memory and cpu profiler left y-axis display
the units in both L1 and L2 (monitoring and usage).

Bug: 189225938
Test: Updated
Change-Id: I2312feb7e617d4b2103f6160801d51f43cb85571