Disable GradleProjectOpenProcessor a little later during startup
GradleProjectOpenProcessor was being disabled using androidStudioInitializer
extension which is invoked from an ApplicationComponent. We want
to remove both application components and the androidStudioInitializer
extension point.

Rather than relying on androidStudioInitializer, we now use the
platform's ApplicationInitializedListener and disable it right after
components are intialized. This CL also renames
`AndroidGradleOpenProjectProcessorInitializer` to
`DisableGradleProjectOpenProcessor` to make it more specific to
how it is used.

Bug: 243082216
Test: manual, launch Studio and verify that GradleProjectOpenProcessor
 is not available during project import.
Change-Id: Id5102d773e77ed82faa8f261e415d8382134d716