Autoupdate of 2 submodules
*tools/adt/idea updated to 8d65c32ec6adc68803bd065170baafa80766f452
Simplify AdbFileProvider & remove early initialization

The logic to find the location of adb relies on AndroidSdkUtils,
and hence it resides in android-core module. In order to keep
android-adb independent of android-core, we currently use a
scheme where application startup and project listeners explicitly
stoer an adb provider into the userdata of an application and
the project respectively.

This CL simplifies this mechanism by utilizing the fact that
we can register application & project services with an interface
defined in android-adb, but implementation provided by core.
This cuts down on the early initialization and the reliance
on userdata fields.

Bug: 200037669
Test: existing tests
Change-Id: I5c4733fa51eaaed7a6023e242e7591448e84c2e0

*tools/vendor/google updated to ed4dbf0203597fea8501557f7ae5298edeeaa528
Register AdbFileProvider & ApkFacetChecker services

See I09d9a9cd1d and I5c4733fa51.

Bug: 200037669
Test: existing tests
Change-Id: Iecc03c5e2d5bceb547edcda78d1248a62221eae4