Add specialized constructs for sequence and conditional
We used to have one `do` notation that subsumed in-sequence,
in-any-order, and conditional, and potentially allowing the assistant
language to inherit more features from Kotlin if needed.
Unfortunately, this approach isn't friendly when we statically know
the steps ahead and want to render them.

This change adds specialized constructs `Seq`, `InAnyOrder`, and `If`,
and removes `Do`.

Bug: n/a
Test: new
Change-Id: I52bf3654e2d9719fdbee03c8eb56f2f8b140c66b