Autoupdate of 4 submodules
*tools/base updated to 12cd17fb5a123c460bed96301952e463feb57107
Fix annotations lookup on a per-platform basis

The code which looks up external annotations corresponding to a
particular platform release used the wrong relative path. However,
there's a fallback where it uses the old annotations from the
platform-tools, which made tests pass until a platform tools update.

Bug: Not filed
Test: ResourceEvaluatorTest
Change-Id: I4d83766f51f87042ac48415d6af14dffc6ab1292

*prebuilts/studio/sdk/darwin updated to 423fc8f2227ac911305133b08b3ffa2c4978ca05
Check in the platform-tools for TiramisuPrivacySandboxSdk.

Test: Covered by existing tests.
Bug: 240533883
Change-Id: I72c016624c4ef55dbaa852a55fda0c53dd1cf0dc

*prebuilts/studio/sdk/linux updated to 7990f39249a4233305dfc1edbd062d56923ed7ce
Check-in  platform-tools for TiramisuPrivacySandbox sdk.

Test: Covered by existing tests.
Bug: 240533883
Change-Id: Id370cd814404679c1fc566eb7f34c08998f34655

*prebuilts/studio/sdk/windows updated to e8edb26d9a673d6b7f3e4c3f2158b76f7eaf30db
Check-in platform-tools for TiramisuPrivacySandbox sdk.

Test: Covered by existing tests.
Bug: 240533883
Change-Id: Icf3cd2283c808d244aa20f7da1f6033b8783d9b6